chore(EMS-3306-3308): no pdf - check your answers - buyer - e2e tests
